feat(server): wire memory.* RPC onto spaceStore (4C-1)
Implement the memory data-plane methods at /api/v1/memory/rpc against the
space store, keeping the wire protocol unchanged and adapting in the handler.

- memory.create/batchCreate/get/update/delete/search/tree/move/deleteTree/
  countTree → spaceStore, gated by the request's treeAccess.
- search mapping: fulltext→bm25, semantic→embed→vec, grep→regexp, tree→ltree,
  meta→metaContains, temporal within/overlaps/contains→space ranges,
  semanticThreshold→maxVecDist (1−t), weights→hybrid weights; fulltext+semantic
  routes to hybridSearch, otherwise search.
- memory.tree builds an lquery from {tree, levels} and reshapes list_tree.
- lossy by design (4C): createdBy is null, search total is the row count,
  orderBy ignored. Postgres errors mapped: 42501→FORBIDDEN, 22023/22P02→
  VALIDATION_ERROR.
- add spaceStore.withTransaction (mirrors coreStore) so batchCreate is atomic.
- fix space search_memory: the filter-only arm emitted `-1 as score` (integer)
  against a float8 return column → `(-1)::float8`.
- integration test covers CRUD, tree/levels, move/deleteTree dryRun, fulltext +
  filter search, grep-alone rejection, and write-access denial.

+++ packages/database/space/migrate/idempotent/002_search.sql
@@ -77,7 +77,8 @@ begin
       );
     end if;
   else
-    _score = $sql$, -1 as score$sql$;
+    -- no ranking arm: constant score, typed float8 to match the return column
+    _score = $sql$, (-1)::float8 as score$sql$;
     _order_by = $sql$order by m.id$sql$;
   end case;
